+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.3.0] - 2026-04-19
+
+### Added
+
+- Added complete non-interactive mode using `DBP_OPERATION` environment variable.
+- Added feature to exclude tables from profiling using `exclude_inventorylist.txt` file.
+
+---
+
 ## [1.2.0] - 2025-01-24
 
 ### Added